+++ src/app/components/system-maintenance/conformance-method/conformance-method-edit.jsx
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import {
   Box,
   Button,
@@ -16,7 +16,6 @@ import * as yup from 'yup';
 import { ChplActionBar } from 'components/action-bar';
 import { ChplTextField } from 'components/util';
 import { sortCriteria } from 'services/criteria.service';
-import { BreadcrumbContext } from 'shared/contexts';
 import {
   criterion as criterionPropType,
   conformanceMethod as conformanceMethodPropType,
@@ -47,14 +46,13 @@ const useStyles = makeStyles({
   },
 });
 
-function ChplConformanceMethodEdit(props) {
-  const {
-    criterionOptions,
-    dispatch,
-    isProcessing,
-    conformanceMethod: initialConformanceMethod,
-  } = props;
-  const { append, display, hide } = useContext(BreadcrumbContext);
+function ChplConformanceMethodEdit({
+  criterionOptions,
+  dispatch,
+  isProcessing,
+  conformanceMethod: initialConformanceMethod,
+  errors: propsErrors = [],
+}) {
   const [criteria, setCriteria] = useState([]);
   const [errors, setErrors] = useState([]);
   const [selectedCriterion, setSelectedCriterion] = useState('');
@@ -63,39 +61,15 @@ function ChplConformanceMethodEdit(props) {
   let formik;
 
   useEffect(() => {
-    append(
-      <Button
-        key="conformanceMethods.add.disabled"
-        depth={2}
-        variant="text"
-        disabled
-      >
-        Add
-      </Button>,
-    );
-    append(
-      <Button
-        key="conformanceMethods.edit.disabled"
-        depth={2}
-        variant="text"
-        disabled
-      >
-        Edit
-      </Button>,
-    );
-  }, []);
-
-  useEffect(() => {
     setConformanceMethod(initialConformanceMethod);
     setCriteria(initialConformanceMethod.criteria?.map((c) => ({
       ...c,
     })) || []);
-    display(initialConformanceMethod.id ? 'conformanceMethods.edit.disabled' : 'conformanceMethods.add.disabled');
   }, [initialConformanceMethod]);
 
   useEffect(() => {
-    setErrors(props.errors.sort((a, b) => (a < b ? -1 : 1))); // eslint-disable-line react/destructuring-assignment
-  }, [props.errors]); // eslint-disable-line react/destructuring-assignment
+    setErrors(propsErrors.sort((a, b) => (a < b ? -1 : 1)));
+  }, [propsErrors]);
 
   const add = (item) => {
     setCriteria((prev) => prev.concat(item));
@@ -113,18 +87,12 @@ function ChplConformanceMethodEdit(props) {
     switch (action) {
       case 'cancel':
         dispatch({ action: 'cancel' });
-        hide('conformanceMethods.add.disabled');
-        hide('conformanceMethods.edit.disabled');
         break;
       case 'delete':
         dispatch({ action: 'delete', payload: buildPayload() });
-        hide('conformanceMethods.add.disabled');
-        hide('conformanceMethods.edit.disabled');
         break;
       case 'save':
         formik.submitForm();
-        hide('conformanceMethods.add.disabled');
-        hide('conformanceMethods.edit.disabled');
         break;
         // no default
     }