+++ src/app/components/system-maintenance/ucd-process/ucd-process-edit.jsx
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import {
   Button,
   makeStyles,
@@ -9,7 +9,6 @@ import * as yup from 'yup';
 
 import { ChplActionBar } from 'components/action-bar';
 import { ChplTextField } from 'components/util';
-import { BreadcrumbContext } from 'shared/contexts';
 import { ucdProcessType } from 'shared/prop-types';
 
 const validationSchema = yup.object({
@@ -31,45 +30,19 @@ const useStyles = makeStyles({
   },
 });
 
-function ChplUcdProcessEdit(props) {
-  const { dispatch, isProcessing } = props;
-  const { append, display, hide } = useContext(BreadcrumbContext);
+function ChplUcdProcessEdit({ dispatch, isProcessing, ucdProcess: initialUcdProcess, errors: propsErrors = [] }) {
   const [errors, setErrors] = useState([]);
   const [ucdProcess, setUcdProcess] = useState({});
   const classes = useStyles();
   let formik;
 
   useEffect(() => {
-    append(
-      <Button
-        key="ucdProcesses.add.disabled"
-        depth={2}
-        variant="text"
-        disabled
-      >
-        Add
-      </Button>,
-    );
-    append(
-      <Button
-        key="ucdProcesses.edit.disabled"
-        depth={2}
-        variant="text"
-        disabled
-      >
-        Edit
-      </Button>,
-    );
-  }, []);
+    setUcdProcess(initialUcdProcess);
+  }, [initialUcdProcess]);
 
   useEffect(() => {
-    setUcdProcess(props.ucdProcess);
-    display(props.ucdProcess.id ? 'ucdProcesses.edit.disabled' : 'ucdProcesses.add.disabled');
-  }, [props.ucdProcess]); // eslint-disable-line react/destructuring-assignment
-
-  useEffect(() => {
-    setErrors(props.errors.sort((a, b) => (a < b ? -1 : 1)));
-  }, [props.errors]); // eslint-disable-line react/destructuring-assignment
+    setErrors(propsErrors.sort((a, b) => (a < b ? -1 : 1)));
+  }, [propsErrors]);
 
   const buildPayload = () => ({
     ...ucdProcess,
@@ -80,18 +53,12 @@ function ChplUcdProcessEdit(props) {
     switch (action) {
       case 'cancel':
         dispatch({ action: 'cancel' });
-        hide('ucdProcesses.add.disabled');
-        hide('ucdProcesses.edit.disabled');
         break;
       case 'delete':
         dispatch({ action: 'delete', payload: buildPayload() });
-        hide('ucdProcesses.add.disabled');
-        hide('ucdProcesses.edit.disabled');
         break;
       case 'save':
         formik.submitForm();
-        hide('ucdProcesses.add.disabled');
-        hide('ucdProcesses.edit.disabled');
         break;
         // no default
     }
@@ -101,7 +68,7 @@ function ChplUcdProcessEdit(props) {
 
   formik = useFormik({
     initialValues: {
-      name: props.ucdProcess?.name || '', // eslint-disable-line react/destructuring-assignment
+      name: initialUcdProcess?.name || '',
     },
     onSubmit: () => {
       dispatch({ action: 'save', payload: buildPayload() });