+++ src/app/components/system-maintenance/ucd-process/ucd-processes.jsx
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import {
   Button,
   Card,
@@ -6,6 +6,7 @@ import {
   CardHeader,
   CircularProgress,
 } from '@material-ui/core';
+import TouchAppOutlinedIcon from '@material-ui/icons/TouchAppOutlined';
 import { useSnackbar } from 'notistack';
 
 import ChplUcdProcessEdit from './ucd-process-edit';
@@ -17,10 +18,8 @@ import {
   usePostUcdProcess,
   usePutUcdProcess,
 } from 'api/standards';
-import { BreadcrumbContext } from 'shared/contexts';
 
 function ChplUcdProcesses() {
-  const { append, display, hide } = useContext(BreadcrumbContext);
   const { data, isLoading, isSuccess } = useFetchUcdProcesses();
   const deleteUcdProcess = useDeleteUcdProcess();
   const postUcdProcess = usePostUcdProcess();
@@ -33,30 +32,6 @@ function ChplUcdProcesses() {
   let handleDispatch;
 
   useEffect(() => {
-    append(
-      <Button
-        key="ucdProcesses.viewall.disabled"
-        depth={1}
-        variant="text"
-        disabled
-      >
-        UCD Processes
-      </Button>,
-    );
-    append(
-      <Button
-        key="ucdProcesses.viewall"
-        depth={1}
-        variant="text"
-        onClick={() => handleDispatch({ action: 'cancel' })}
-      >
-        UCD Processes
-      </Button>,
-    );
-    display('ucdProcesses.viewall.disabled');
-  }, []);
-
-  useEffect(() => {
     if (isLoading || !isSuccess) { return; }
     setUcdProcesses(data);
   }, [data, isLoading, isSuccess]);
@@ -66,10 +41,6 @@ function ChplUcdProcesses() {
       case 'cancel':
         setActiveUcdProcess(undefined);
         setIsProcessing(false);
-        display('ucdProcesses.viewall.disabled');
-        hide('ucdProcesses.viewall');
-        hide('ucdProcesses.add.disabled');
-        hide('ucdProcesses.edit.disabled');
         break;
       case 'delete':
         setErrors([]);
@@ -81,8 +52,6 @@ function ChplUcdProcesses() {
             });
             setActiveUcdProcess(undefined);
             setIsProcessing(false);
-            display('ucdProcesses.viewall.disabled');
-            hide('ucdProcesses.viewall');
           },
           onError: (error) => {
             setErrors(error.response.data.errorMessages);
@@ -93,8 +62,6 @@ function ChplUcdProcesses() {
       case 'edit':
         setActiveUcdProcess(payload);
         setErrors([]);
-        display('ucdProcesses.viewall');
-        hide('ucdProcesses.viewall.disabled');
         break;
       case 'save':
         setErrors([]);
@@ -107,8 +74,6 @@ function ChplUcdProcesses() {
               });
               setActiveUcdProcess(undefined);
               setIsProcessing(false);
-              display('ucdProcesses.viewall.disabled');
-              hide('ucdProcesses.viewall');
             },
             onError: (error) => {
               setErrors(error.response.data.errorMessages);
@@ -123,8 +88,6 @@ function ChplUcdProcesses() {
               });
               setActiveUcdProcess(undefined);
               setIsProcessing(false);
-              display('ucdProcesses.viewall.disabled');
-              hide('ucdProcesses.viewall');
             },
             onError: (error) => {
               setErrors(error.response.data?.errorMessages);
@@ -140,7 +103,12 @@ function ChplUcdProcesses() {
   if (activeUcdProcess) {
     return (
       <Card>
-        <CardHeader title={`${activeUcdProcess.id ? 'Edit' : 'Add'} UCD Process`} />
+        <CardHeader title={(
+          <>
+            <TouchAppOutlinedIcon style={{ verticalAlign: 'middle', marginRight: '8px' }} />
+            {`${activeUcdProcess.id ? 'Edit' : 'Add'} UCD Process`}
+          </>
+)} />
         <CardContent>
           <ChplUcdProcessEdit
             ucdProcess={activeUcdProcess}
@@ -161,7 +129,15 @@ function ChplUcdProcesses() {
 
   return (
     <Card>
-      <CardHeader title="UCD Processes" />
+      <CardHeader
+        style={{ paddingLeft: '16px' }}
+        title={(
+          <>
+            UCD Processes
+            <TouchAppOutlinedIcon style={{ verticalAlign: 'middle', marginLeft: '8px' }} />
+          </>
+)}
+      />
       <CardContent>
         { (deleteUcdProcess.isLoading || postUcdProcess.isLoading || putUcdProcess.isLoading)
           && (