+++ src/app/components/util/index.js
@@ -8,7 +8,9 @@ import ChplNonProdIndicator from './non-prod-indicator';
 import ChplPagination from './pagination';
 import ChplProgress from './progress';
 import ChplLoadingTable from './loading-table';
+import ChplSortControls from './chpl-sort-controls';
 import { ChplSortableHeaders } from './sortable-headers';
+import ChplSearchResultCard from './chpl-search-result-card';
 import ChplTextField from './chpl-text-field';
 import ChplTooltip from './chpl-tooltip';
 import ChplUpdateIndicator from './update-indicator';
@@ -26,6 +28,8 @@ export {
   ChplNonProdIndicator,
   ChplPagination,
   ChplProgress,
+  ChplSearchResultCard,
+  ChplSortControls,
   ChplSortableHeaders,
   ChplTextField,
   ChplTooltip,